+++ R/utils.R
@@ -78,9 +78,9 @@ mclock <- function() .Call(rnng_clock)
 #'
 #' @return Invisible NULL.
 #'
-#' @details Non-integer values for \sQuote{time} are coerced to integer, and the
-#'     absolute value is taken (the sign is ignored). Non-numeric values are
-#'     ignored, causing the function to return immediately.
+#' @details Non-integer values for \sQuote{time} are coerced to integer.
+#'     Negative / NA / non-numeric values are ignored, causing the function to
+#'     return immediately.
 #'
 #'     Note that unlike \code{\link{Sys.sleep}}, this function is not
 #'     user-interruptible by sending SIGINT e.g. with ctrl + c.