+++ src/nanonext.h
@@ -203,7 +203,7 @@ typedef struct nano_thread_duo_s {
   (x)->len = 0;                                                \
   (x)->cur = sz
 #define NANO_FREE(x) if (x.len) R_Free(x.buf)
-#define NANO_INTEGER(x) (int *) DATAPTR_RO(x)
+#define NANO_INTEGER(x) *(int *) DATAPTR_RO(x)
 #define NANO_ERROR(x) { Rf_error(x); return R_NilValue; }
 #define NANO_CLASS2(x, cls1, cls2)                             \
   SEXP klass = Rf_allocVector(STRSXP, 2);                      \