+++ src/utils.c
@@ -96,12 +96,16 @@ SEXP rnng_clock(void) {
 
 SEXP rnng_sleep(SEXP msec) {
 
+  int time;
   switch (TYPEOF(msec)) {
   case INTSXP:
-    nng_msleep((nng_duration) abs(INTEGER(msec)[0]));
+  case LGLSXP:
+    time = NANO_INTEGER(msec);
+    if (time > 0) nng_msleep((nng_duration) time);
     break;
   case REALSXP:
-    nng_msleep((nng_duration) abs(Rf_asInteger(msec)));
+    time = Rf_asInteger(msec);
+    if (time > 0) nng_msleep((nng_duration) time);
     break;
   }
 
@@ -241,7 +245,7 @@ SEXP rnng_ncurl(SEXP http, SEXP convert, SEXP follow, SEXP method, SEXP headers,
 
   code = nng_http_res_get_status(res), relo = code >= 300 && code < 400;
 
-  if (relo && *NANO_INTEGER(follow)) {
+  if (relo && NANO_INTEGER(follow)) {
     const char *location = nng_http_res_get_header(res, "Location");
     if (location == NULL) goto resume;
     nng_url *oldurl = url;
@@ -294,7 +298,7 @@ SEXP rnng_ncurl(SEXP http, SEXP convert, SEXP follow, SEXP method, SEXP headers,
 
   nng_http_res_get_data(res, &dat, &sz);
 
-  if (*NANO_INTEGER(convert)) {
+  if (NANO_INTEGER(convert)) {
     vec = rawToChar(dat, sz);
   } else {
     vec = Rf_allocVector(RAWSXP, sz);
@@ -338,7 +342,7 @@ SEXP rnng_stream_dial(SEXP url, SEXP textframes, SEXP tls) {
     Rf_error("'tls' is not a valid TLS Configuration");
   nano_stream *nst = R_Calloc(1, nano_stream);
   nst->mode = NANO_STREAM_DIALER;
-  nst->textframes = *NANO_INTEGER(textframes) != 0;
+  nst->textframes = NANO_INTEGER(textframes) != 0;
   nst->tls = NULL;
   nng_url *up;
   nng_aio *aiop;
@@ -427,7 +431,7 @@ SEXP rnng_stream_listen(SEXP url, SEXP textframes, SEXP tls) {
     Rf_error("'tls' is not a valid TLS Configuration");
   nano_stream *nst = R_Calloc(1, nano_stream);
   nst->mode = NANO_STREAM_LISTENER;
-  nst->textframes = *NANO_INTEGER(textframes) != 0;
+  nst->textframes = NANO_INTEGER(textframes) != 0;
   nst->tls = NULL;
   nng_url *up;
   nng_aio *aiop;
@@ -608,7 +612,7 @@ SEXP rnng_status_code(SEXP x) {
 
 SEXP rnng_tls_config(SEXP client, SEXP server, SEXP pass, SEXP auth) {
 
-  const nng_tls_auth_mode mod = *NANO_INTEGER(auth) ? NNG_TLS_AUTH_MODE_REQUIRED : NNG_TLS_AUTH_MODE_OPTIONAL;
+  const nng_tls_auth_mode mod = NANO_INTEGER(auth) ? NNG_TLS_AUTH_MODE_REQUIRED : NNG_TLS_AUTH_MODE_OPTIONAL;
   R_xlen_t usefile;
   nng_tls_config *cfg;
   int xc;