+++ NEWS.md
@@ -10,6 +10,7 @@
 * `request()` specifying argument 'cv' other than NULL or a 'conditionVariable' will cause the pipe connection to be dropped when the reply is (asynchronously) completed. 
 * Removes deprecated functions `strcat()`, `recv_aio_signal()` and `request_signal()`.
 * Drops `base64enc()` and `base64dec()` in favour of those from the {secretbase} package.
+* `msleep()` now ignores negative values rather than taking the absolute value.
 * `later` is now relaxed to a soft 'suggests' dependency (only required if using promises).
 * `promises` is added as a soft 'enhances' dependency.